Ensure delays from simulation are shown in live situation
